Pushes test population into a blank CDM.

Pushes test population into a blank CDM.

Usage

patientsCDM(
  pathJson = NULL,
  testName = NULL,
  cdmVersion = "5.3",
  cdmName = NULL
)

Arguments

pathJson

Directory where the sample populations in json are located. If NULL, gets the default inst/testCases directory.

testName

Name of the sample population JSON file. If NULL it will push the first sample population in the testCases directory.

cdmVersion

cdm version, default "5.3".

cdmName

Name of the cdm, default NULL.

Value

A CDM reference object with a sample population.

Examples

# \donttest{
filePath <- system.file("extdata", "testPatientsRSV.xlsx", package = "TestGenerator")
TestGenerator::readPatients(filePath = filePath, outputPath = tempdir())
#> ✔ All tables are valid
#> ✔ Unit Test Definition Created Successfully: 'test'
cdm <- TestGenerator::patientsCDM(pathJson = tempdir(), testName = "test")
#> 
#> Download completed!
#> Creating CDM database /tmp/RtmpiMJMiU/empty_cdm_5.3.zip
#> ■■■■■■■■■■■■■■■■■■■■              63% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■          76% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | ETA:  4s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       85% | ETA:  3s
#> ! cdm name not specified and could not be inferred from the cdm source table
#> ✖ Standard table(s) in test data: person, observation_period, condition_occurrence, visit_occurrence, visit_detail and death
#> ✔ Patients pushed to blank CDM successfully
duckdb::duckdb_shutdown(duckdb::duckdb())
# }